+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 Change:
 - conditional parseQuery parameter
 - conditional pass `c.request` to handler for streaming response
+- fix missing `contentType` type on `parser`
 
 # 1.4.16 - 13 Nov 2025
 Improvement: